+++ src/aux-common/Files/FileCalculations.ts
@@ -16,6 +16,7 @@ import {
     FileLabelAnchor,
     DEFAULT_LABEL_ANCHOR,
     FileDragMode,
+    ContextVisualizeMode,
 } from './File';
 
 import uuid from 'uuid/v4';
@@ -995,7 +996,7 @@ export function createWorkspace(
             'aux.context.surface.x': 0,
             'aux.context.surface.y': 0,
             'aux.context.surface.z': 0,
-            'aux.context.surface': true,
+            'aux.context.visualize': 'surface',
             'aux.context.locked': locked,
             'aux.context': builderContextId,
         },
@@ -1611,7 +1612,12 @@ export function getFileDragMode(
     if (typeof val === 'boolean') {
         return val ? 'all' : 'none';
     }
-    if (val === 'clone' || val === 'pickup' || val === 'drag') {
+    if (
+        val === 'clone' ||
+        val === 'pickup' ||
+        val === 'drag' ||
+        val === 'diff'
+    ) {
         return val;
     } else {
         return 'all';
@@ -1724,25 +1730,35 @@ export function getContextSize(
     calc: FileCalculationContext,
     contextFile: File
 ): number {
-    return calculateNumericalTagValue(
-        calc,
-        contextFile,
-        `aux.context.surface.size`,
-        isUserFile(contextFile) ? 0 : DEFAULT_WORKSPACE_SIZE
-    );
+    if (getContextVisualizeMode(calc, contextFile) === 'surface') {
+        return calculateNumericalTagValue(
+            calc,
+            contextFile,
+            `aux.context.surface.size`,
+            DEFAULT_WORKSPACE_SIZE
+        );
+    }
+    return 0;
 }
 
 /**
- * Determines if a context's surface should be visible.
- * Defaults to false.
- * @param calc The file calculation context.
+ * Gets the aux.context.visualize mode from the given file.
+ * @param calc The calculation context.
  * @param file The file.
  */
-export function isContextSurfaceVisible(
+export function getContextVisualizeMode(
     calc: FileCalculationContext,
     file: File
-): boolean {
-    return calculateBooleanTagValue(calc, file, 'aux.context.surface', false);
+): ContextVisualizeMode {
+    const val = calculateFileValue(calc, file, 'aux.context.visualize');
+    if (typeof val === 'boolean') {
+        return val;
+    }
+    if (val === 'surface') {
+        return val;
+    } else {
+        return false;
+    }
 }
 
 /**
@@ -1932,21 +1948,28 @@ export function isPickupable(
 /**
  * Gets a partial file that can be used to apply the diff that the given file represents.
  * A diff file is any file that has `aux._diff` set to `true` and `aux._diffTags` set to a list of tag names.
+ * @param calc The file calculation context.
  * @param file The file that represents the diff.
  */
-export function getDiffUpdate(file: File): PartialFile {
+export function getDiffUpdate(
+    calc: FileCalculationContext,
+    file: File
+): PartialFile {
     if (isDiff(file)) {
         let update: PartialFile = {
             tags: {},
         };
 
         let tags = tagsOnFile(file);
-        let diffTags = file.tags['aux._diffTags'];
+        let diffTags =
+            calculateFileValue(calc, file, 'aux.movable.diffTags') ||
+            file.tags['aux._diffTags'];
         for (let i = 0; i < tags.length; i++) {
             let tag = tags[i];
             if (
                 tag === 'aux._diff' ||
                 tag === 'aux._diffTags' ||
+                tag === 'aux.movable.diffTags' ||
                 diffTags.indexOf(tag) < 0
             ) {
                 continue;