+++ src/aux-common/Formulas/formula-lib.ts
@@ -18,6 +18,8 @@ import {
     showQRCode as calcShowQRCode,
     goToContext as calcGoToContext,
     importAUX as calcImportAUX,
+    showInputForTag as calcShowInputForTag,
+    ShowInputOptions,
 } from '../Files/FilesChannel';
 import uuid from 'uuid/v4';
 import { every, find, sortBy } from 'lodash';
@@ -277,7 +279,10 @@ export function destroy(file: FileProxy | string | FileProxy[]) {
  * @param file The file, file ID, or list of files to destroy the tag sections of.
  * @param tagSection The tag section to remove on the file.
  */
-export function removeTags(file: FileProxy | FileProxy[], tagSection: string) {
+export function removeTags(
+    file: FileProxy | FileProxy[],
+    tagSection: string | RegExp
+) {
     if (typeof file === 'object' && Array.isArray(file)) {
         let fileList: any[] = file;
 
@@ -285,7 +290,11 @@ export function removeTags(file: FileProxy | FileProxy[], tagSection: string) {
             let tags = tagsOnFile(fileList[h]);
 
             for (let i = tags.length - 1; i >= 0; i--) {
-                if (tags[i].includes(tagSection)) {
+                if (tagSection instanceof RegExp) {
+                    if (tagSection.test(tags[i])) {
+                        fileList[h][tags[i]] = null;
+                    }
+                } else if (tags[i].includes(tagSection)) {
                     let doRemoveTag = false;
 
                     if (tags[i].includes('.')) {
@@ -309,7 +318,11 @@ export function removeTags(file: FileProxy | FileProxy[], tagSection: string) {
 
         for (let i = tags.length - 1; i >= 0; i--) {
             // if the tag section is relevant to the curretn tag at all
-            if (tags[i].includes(tagSection)) {
+            if (tagSection instanceof RegExp) {
+                if (tagSection.test(tags[i])) {
+                    file[tags[i]] = null;
+                }
+            } else if (tags[i].includes(tagSection)) {
                 let doRemoveTag = false;
                 // if this tag has a period in it, check for first word to match
                 if (tags[i].includes('.')) {
@@ -515,6 +528,15 @@ function whisper(
  */
 function goToContext(context: string) {
     actions.push(calcGoToContext(context));
+}
+
+function showInputForTag(
+    file: FileProxy | string,
+    tag: string,
+    options?: Partial<ShowInputOptions>
+) {
+    const id = typeof file === 'string' ? file : file.id;
+    actions.push(calcShowInputForTag(id, tag, options));
 }
 
 /**
@@ -1000,6 +1022,7 @@ export const player = {
     isConnected,
     currentContext,
     isBuilder,
+    showInputForTag,
 };
 
 /**