+++ src/aux-server/aux-web/aux-player/PlayerApp/PlayerApp.ts
@@ -20,6 +20,16 @@ import {
     calculateDestroyFileEvents,
     merge,
     simulationIdToString,
+    FileCalculationContext,
+    AuxFile,
+    calculateFileValue,
+    calculateFormattedFileValue,
+    getFileInputTarget,
+    getFileInputPlaceholder,
+    ShowInputForTagEvent,
+    ShowInputOptions,
+    ShowInputType,
+    ShowInputSubtype,
 } from '@casual-simulation/aux-common';
 import SnackbarOptions from '../../shared/SnackbarOptions';
 import { copyToClipboard } from '../../shared/SharedUtils';
@@ -31,12 +41,24 @@ import HexIcon from '../public/icons/Hexagon.svg';
 import { QrcodeStream } from 'vue-qrcode-reader';
 import { Simulation } from '../../shared/Simulation';
 import { SidebarItem } from '../../shared/vue-components/BaseGameView';
+import { Swatches, Chrome, Compact } from 'vue-color';
+
+export interface SidebarItem {
+    id: string;
+    group: string;
+    text: string;
+    icon: string;
+    click: () => void;
+}
 
 @Component({
     components: {
         app: PlayerApp,
         'qr-code': QRCode,
         'qrcode-stream': QrcodeStream,
+        'color-picker-swatches': Swatches,
+        'color-picker-advanced': Chrome,
+        'color-picker-basic': Compact,
     },
 })
 export default class PlayerApp extends Vue {
@@ -108,6 +130,18 @@ export default class PlayerApp extends Vue {
      * The QR Code to show.
      */
     qrCode: string = '';
+
+    inputDialogLabel: string = '';
+    inputDialogPlaceholder: string = '';
+    inputDialogInput: string = '';
+    inputDialogType: ShowInputType = 'text';
+    inputDialogSubtype: ShowInputSubtype = 'basic';
+    inputDialogInputValue: any = '';
+    inputDialogTarget: AuxFile = null;
+    inputDialogLabelColor: string = '#000';
+    inputDialogBackgroundColor: string = '#FFF';
+    showInputDialog: boolean = false;
+    inputDialogSimulation: Simulation = null;
 
     confirmDialogOptions: ConfirmDialogOptions = new ConfirmDialogOptions();
     alertDialogOptions: AlertDialogOptions = new AlertDialogOptions();
@@ -425,6 +459,8 @@ export default class PlayerApp extends Vue {
                     });
 
                     this._updateQuery();
+                } else if (e.name === 'show_input_for_tag') {
+                    this._showInputDialog(simulation, e);
                 }
             }),
             simulation.aux.channel.connectionStateChanged.subscribe(
@@ -457,6 +493,114 @@ export default class PlayerApp extends Vue {
         this.simulations.push(info);
 
         this._updateQuery();
+    }
+
+    // TODO: Move to a shared class/component
+    _showInputDialog(simulation: Simulation, event: ShowInputForTagEvent) {
+        const calc = simulation.helper.createContext();
+        const file = simulation.helper.filesState[event.fileId];
+        this._updateLabel(calc, file, event.tag, event.options);
+        this._updateColor(calc, file, event.options);
+        this._updateInput(calc, file, event.tag, event.options);
+        this.inputDialogSimulation = simulation;
+        this.showInputDialog = true;
+    }
+
+    updateInputDialogColor(newColor: any) {
+        if (typeof newColor === 'object') {
+            this.inputDialogInputValue = newColor.hex;
+        } else {
+            this.inputDialogInputValue = newColor;
+        }
+    }
+
+    async closeInputDialog() {
+        if (this.showInputDialog) {
+            await this.inputDialogSimulation.helper.action('onClose', [
+                this.inputDialogTarget,
+            ]);
+            this.showInputDialog = false;
+        }
+    }
+
+    async saveInputDialog() {
+        if (this.showInputDialog) {
+            let value: any;
+            if (
+                this.inputDialogType === 'color' &&
+                typeof this.inputDialogInputValue === 'object'
+            ) {
+                value = this.inputDialogInputValue.hex;
+            } else {
+                value = this.inputDialogInputValue;
+            }
+            await this.inputDialogSimulation.helper.updateFile(
+                this.inputDialogTarget,
+                {
+                    tags: {
+                        [this.inputDialogInput]: value,
+                    },
+                }
+            );
+            await this.inputDialogSimulation.helper.action('onSave', [
+                this.inputDialogTarget,
+            ]);
+            await this.closeInputDialog();
+        }
+    }
+
+    private _updateColor(
+        calc: FileCalculationContext,
+        file: AuxFile,
+        options: Partial<ShowInputOptions>
+    ) {
+        if (typeof options.backgroundColor !== 'undefined') {
+            this.inputDialogBackgroundColor = options.backgroundColor;
+        } else {
+            this.inputDialogBackgroundColor = '#FFF';
+        }
+    }
+
+    private _updateLabel(
+        calc: FileCalculationContext,
+        file: AuxFile,
+        tag: string,
+        options: Partial<ShowInputOptions>
+    ) {
+        if (typeof options.title !== 'undefined') {
+            this.inputDialogLabel = options.title;
+        } else {
+            this.inputDialogLabel = tag;
+        }
+
+        if (typeof options.foregroundColor !== 'undefined') {
+            this.inputDialogLabelColor = options.foregroundColor;
+        } else {
+            this.inputDialogLabelColor = '#000';
+        }
+    }
+
+    private _updateInput(
+        calc: FileCalculationContext,
+        file: AuxFile,
+        tag: string,
+        options: Partial<ShowInputOptions>
+    ) {
+        this.inputDialogInput = tag;
+        this.inputDialogType = options.type || 'text';
+        this.inputDialogSubtype = options.subtype || 'basic';
+        this.inputDialogTarget = file;
+        this.inputDialogInputValue = calculateFormattedFileValue(
+            calc,
+            this.inputDialogTarget,
+            this.inputDialogInput
+        );
+
+        if (typeof options.placeholder !== 'undefined') {
+            this.inputDialogPlaceholder = options.placeholder;
+        } else {
+            this.inputDialogPlaceholder = this.inputDialogInput;
+        }
     }
 
     private _simulationRemoved(simulation: Simulation) {