+++ src/aux-server/aux-web/aux-projector/BuilderGameView/BuilderGameView.vue
@@ -34,18 +34,6 @@
                     id="vr-button-container"
                     class="vr-button-container"
                 ></span>
-
-                <div ref="fileQueue">
-                    <div v-if="!vrDisplay && filesMode && simulation3D" class="toolbar-layout">
-                        <mini-file
-                            v-for="(file, index) in simulation3D.recentFiles"
-                            :key="index"
-                            :file="file"
-                            :selected="simulation3D.selectedRecentFile === file"
-                            :large="index === 0"
-                        ></mini-file>
-                    </div>
-                </div>
             </div>
             <trash-can v-if="showTrashCan" ref="trashCan"></trash-can>
             <div class="toolbar right">