+++ src/aux-server/aux-web/shared/FileManager.ts
@@ -438,7 +438,7 @@ export class FileManager implements Simulation {
             await this.helper.createFile(this._appManager.user.id, {
                 [userContext]: true,
                 ['aux.context']: userContext,
-                ['aux.context.surface']: true,
+                ['aux.context.visualize']: true,
                 ['aux._user']: this._appManager.user.username,
                 ['aux._userInventoryContext']: userInventoryContext,
                 ['aux._userMenuContext']: userMenuContext,