+++ src/aux-server/server/projector.config.ts
@@ -1,10 +1,12 @@
 import { ClientConfig } from './config';
+import * as process from 'process';
 
 const config: ClientConfig = {
     index: 'projector.html',
     web: {
         isBuilder: true,
         isPlayer: false,
+        sentryDsn: process.env.SENTRY_DSN,
     },
 };